refactor: 💡 add type to notifications
