+++ src/resources/scripts/helpers/callHelpers.ts
@@ -1,4 +1,5 @@
 export type CallType = 'video' | 'audio';
+export type NotificationType = 'call';
 
 export const getCallUrl = (callType: CallType, url: string) => {
 	return callType === 'video'
